% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{initialize_app_config}
\alias{initialize_app_config}
\title{Initialize app configuration and tokens}
\usage{
initialize_app_config()
}
\value{
A list containing API URL and tokens
}
\description{
Initialize app configuration and tokens
}
