% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_data_dictionaries}
\alias{load_data_dictionaries}
\title{Load data dictionaries from REDCap}
\usage{
load_data_dictionaries(config)
}
\arguments{
\item{config}{The configuration list from initialize_app_config}
}
\value{
A list containing data dictionaries
}
\description{
Load data dictionaries from REDCap
}
