% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_milestone_data}
\alias{load_milestone_data}
\title{Load milestone data from REDCap}
\usage{
load_milestone_data(config)
}
\arguments{
\item{config}{The configuration list from initialize_app_config}
}
\value{
A list containing processed milestone data
}
\description{
Load milestone data from REDCap
}
