% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_resident_data}
\alias{load_resident_data}
\title{Load resident data from REDCap}
\usage{
load_resident_data(config)
}
\arguments{
\item{config}{The configuration list from initialize_app_config}
}
\value{
Processed resident data
}
\description{
Load resident data from REDCap
}
