% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{render_next_steps_block}
\alias{render_next_steps_block}
\title{Render the “Next Steps” block for the Graduating section}
\usage{
render_next_steps_block(rdm_dict, graduating_fields)
}
\arguments{
\item{rdm_dict}{A data frame of your REDCap dictionary}

\item{graduating_fields}{A character vector of field_names to include for Graduating}
}
\value{
A tagList of UI elements
}
\description{
Render the “Next Steps” block for the Graduating section
}
